requirements.txt
ipykernel
pygame
psycopg2-binary
python-dotenv
SQLAlchemy
sqlalchemy-utils
